Former-commit-id: e3382f5af462f5146fde5e137009291afa91140d

+++ backend/src/main/java/com/ssafy/i5i/hotelAPI/domain/hotel/dto/request/AttractionNameRequestDto.java
@@ -8,6 +8,6 @@ import lombok.*;
 @AllArgsConstructor
 public class AttractionNameRequestDto {
     private String attractionName;
-    private Long distance;
+    private Double distance;
     private String sorted;
 }
